+++ common/src/main/java/net/xstopho/resourceconfigapi/client/gui/screen/ResourceConfigScreen.java
@@ -90,7 +90,7 @@ public class ResourceConfigScreen extends Screen {
 
         this.repositionElements();
 
-        ComponentUtils.printMissingTranslations(modId);
+        ComponentUtils.logMissingTranslations(modId);
     }
 
     @Override