+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### v3.3.1
 - fixing an issue where undo button was falsely active
+- fixing logging for missing translation keys
+-   - outputs the usable translation keys for your config
 - adding Indicator for Config Option if a world/game restart is necessary
 - adding more translations